Use ephemeral browser in iOS >=13
This will not share the browser state inside the OpenID popup with the
regular Safari instance. It has the downside that the user has to
re-login even if they are already logged in to Contentpass in the
regular browser. But it has the big benefit that we get rid of the
confirmation dialog before the OpenID popup.